Refactor things into separate files
